requirements.txt
python-telegram-bot==21.6
APScheduler==3.10.4
pydantic==2.9.2
pydantic-settings==2.6.1
MetaTrader5==5.0.4500; sys_platform == "win32"
